Messages had a small error where someone accepting the offer to exit a contract early would *say* the land lord paid, when the tennant was paying
Removed CoreProtect as a soft-depend as it does not appear to actually
do what I need.

Updated the "Self cleaning" rentals so that broken blocks do not restore
themselves upon breaking the rental agreement. The plugin will restore
*only* air, so new structures will be destroyed, unless they occupy a
block that was already present (replacing a dirt block with a chest will
result in the chest remaining when the agreement breaks).